leetcode/leetcode-04/src/main/java/Solution376.java
public class Solution376 {
    public int wiggleMaxLength(int[] nums) {
        int len = nums.length;
        if (len < 2) {
            return len;
        }
        // up[i] 表示以前 i 个元素中的某一个为结尾的最长的，最后两项差值为正数，如 [1,3,2,4]
        int[] up = new int[len];
        // down[i] 表示以前 i 个元素中的某一个为结尾的最长的，最后两项差值为负数，如 [4,2,3,1]
        int[] down = new int[len];
        // 初始状态
        up[0] = 1;
        down[0] = 1;
        // 状态转移
        for (int i = 1; i < len; i++) {
            if (nums[i] > nums[i - 1]) {
                up[i] = Math.max(up[i - 1], down[i - 1] + 1);
                down[i] = down[i - 1];
            } else if (nums[i] < nums[i - 1]) {
                up[i] = up[i - 1];
                down[i] = Math.max(up[i - 1] + 1, down[i - 1]);
            } else {
                up[i] = up[i - 1];
                down[i] = down[i - 1];
            }
        }
        return Math.max(up[len - 1], down[len - 1]);
    }
}
/*
376. 摆动序列
https://leetcode.cn/problems/wiggle-subsequence/

如果连续数字之间的差严格地在正数和负数之间交替，则数字序列称为 摆动序列 。第一个差（如果存在的话）可能是正数或负数。
仅有一个元素或者含两个不等元素的序列也视作摆动序列。
例如， [1, 7, 4, 9, 2, 5] 是一个 摆动序列 ，因为差值 (6, -3, 5, -7, 3) 是正负交替出现的。
相反，[1, 4, 7, 2, 5] 和 [1, 7, 4, 5, 5] 不是摆动序列，第一个序列是因为它的前两个差值都是正数，第二个序列是因为它的最后一个差值为零。
子序列 可以通过从原始序列中删除一些（也可以不删除）元素来获得，剩下的元素保持其原始顺序。
给你一个整数数组 nums ，返回 nums 中作为 摆动序列 的 最长子序列的长度 。
提示：
1 <= nums.length <= 1000
0 <= nums[i] <= 1000
进阶：你能否用 O(n) 时间复杂度完成此题?

动态规划。
 */